Finished Project
Everything completed. Sort is done in sorting_algo.vhd. UART works as
expected, in all regards. Also added a python script to show the output
in a more readable format. See "output_script.py". DATA_WIDTH in this
script refers to number of hex values. So if 5 hex values per line,
DATA_WIDTH = 5 for the script.